@@ -1,17 +1,19 @@
 package org.acheron.authserver.repo;
 
+import org.acheron.authserver.dto.UserDto;
 import org.acheron.authserver.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-    Optional<User> findByUsername(String username);
+public interface UserRepository extends CrudRepository<User,Long> {
+    User findByUsername(String username);
+
+    boolean existsByUsername(String username);
 
     boolean existsByEmail(String email);
 
-    boolean existsByUsername(String username);
+    Optional<User> findByEmail(String email);
 }
