@@ -1,10 +1,9 @@
-package org.acheron.authserver.entity;
+package com.acheron.userserver.entity;
 
-import jakarta.persistence.*;
-import lombok.*;
 import org.hibernate.proxy.HibernateProxy;
 import org.springframework.format.annotation.DateTimeFormat;
-
+import jakarta.persistence.*;
+import lombok.*;
 import java.time.LocalDateTime;
 import java.util.Objects;
 
@@ -49,7 +48,7 @@ public class Token {
         Class<?> oEffectiveClass = o instanceof HibernateProxy ? ((HibernateProxy) o).getHibernateLazyInitializer().getPersistentClass() : o.getClass();
         Class<?> thisEffectiveClass = this instanceof HibernateProxy ? ((HibernateProxy) this).getHibernateLazyInitializer().getPersistentClass() : this.getClass();
         if (thisEffectiveClass != oEffectiveClass) return false;
-        org.acheron.authserver.entity.Token that = (org.acheron.authserver.entity.Token) o;
+        com.acheron.userserver.entity.Token that = (com.acheron.userserver.entity.Token) o;
         return getId() != null && Objects.equals(getId(), that.getId());
     }
 
