@@ -1,6 +1,6 @@
-package org.acheron.authserver.repo;
+package com.acheron.userserver.repo;
 
-import org.acheron.authserver.entity.Token;
+import com.acheron.userserver.entity.Token;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
